Extract `sleep_or_wait()` to async primitives
