+++ kopf/reactor/activities.py
@@ -19,9 +19,10 @@ The process is intentionally split into multiple packages:
 import logging
 from typing import Mapping, MutableMapping, NoReturn
 
-from kopf.reactor import causation, effects, handling, lifecycles, registries
+from kopf.reactor import causation, handling, lifecycles, registries
 from kopf.storage import states
-from kopf.structs import callbacks, configuration, credentials, handlers as handlers_, memos
+from kopf.structs import callbacks, configuration, credentials, \
+                         handlers as handlers_, memos, primitives
 
 logger = logging.getLogger(__name__)
 
@@ -123,7 +124,7 @@ async def run_activity(
         )
         outcomes.update(current_outcomes)
         state = state.with_outcomes(current_outcomes)
-        await effects.sleep_or_wait(state.delay)
+        await primitives.sleep_or_wait(state.delay)
 
     # Activities assume that all handlers must eventually succeed.
     # We raise from the 1st exception only: just to have something real in the tracebacks.
