+++ kopf/reactor/daemons.py
@@ -413,7 +413,7 @@ async def _resource_daemon(
     body = cause.body
 
     if handler.initial_delay is not None:
-        await effects.sleep_or_wait(handler.initial_delay, cause.stopper)
+        await primitives.sleep_or_wait(handler.initial_delay, cause.stopper)
 
     # Similar to activities (in-memory execution), but applies patches on every attempt.
     state = states.State.from_scratch().with_handlers([handler])
@@ -433,7 +433,7 @@ async def _resource_daemon(
 
         # The in-memory sleep does not react to resource changes, but only to stopping.
         if state.delay:
-            await effects.sleep_or_wait(state.delay, cause.stopper)
+            await primitives.sleep_or_wait(state.delay, cause.stopper)
 
     if stopper.is_set():
         logger.debug(f"{handler} has exited on request and will not be retried or restarted.")
@@ -476,7 +476,7 @@ async def _resource_timer(
     body = cause.body
 
     if handler.initial_delay is not None:
-        await effects.sleep_or_wait(handler.initial_delay, stopper)
+        await primitives.sleep_or_wait(handler.initial_delay, stopper)
 
     # Similar to activities (in-memory execution), but applies patches on every attempt.
     state = states.State.from_scratch().with_handlers([handler])
@@ -492,7 +492,7 @@ async def _resource_timer(
         if handler.idle is not None:
             while not stopper.is_set() and time.monotonic() - memory.idle_reset_time < handler.idle:
                 delay = memory.idle_reset_time + handler.idle - time.monotonic()
-                await effects.sleep_or_wait(delay, stopper)
+                await primitives.sleep_or_wait(delay, stopper)
             if stopper.is_set():
                 continue
 
@@ -515,7 +515,7 @@ async def _resource_timer(
         # For temporary errors, override the schedule by the one provided by errors themselves.
         # It can be either a delay from TemporaryError, or a backoff for an arbitrary exception.
         if not state.done:
-            await effects.sleep_or_wait(state.delays, stopper)
+            await primitives.sleep_or_wait(state.delays, stopper)
 
         # For sharp timers, calculate how much time is left to fit the interval grid:
         #       |-----|-----|-----|-----|-----|-----|---> (interval=5, sharp=True)
@@ -523,19 +523,19 @@ async def _resource_timer(
         elif handler.interval is not None and handler.sharp:
             passed_duration = time.monotonic() - started
             remaining_delay = handler.interval - (passed_duration % handler.interval)
-            await effects.sleep_or_wait(remaining_delay, stopper)
+            await primitives.sleep_or_wait(remaining_delay, stopper)
 
         # For regular (non-sharp) timers, simply sleep from last exit to the next call:
         #       |-----|-----|-----|-----|-----|-----|---> (interval=5, sharp=False)
         #       [slow_handler].....[slow_handler].....[slow...
         elif handler.interval is not None:
-            await effects.sleep_or_wait(handler.interval, stopper)
+            await primitives.sleep_or_wait(handler.interval, stopper)
 
         # For idle-only no-interval timers, wait till the next change (i.e. idling reset).
         # NB: This will skip the handler in the same tact (1/64th of a second) even if changed.
         elif handler.idle is not None:
             while memory.idle_reset_time <= started:
-                await effects.sleep_or_wait(handler.idle, stopper)
+                await primitives.sleep_or_wait(handler.idle, stopper)
 
         # Only in case there are no intervals and idling, treat it as a one-shot handler.
         # This makes the handler practically meaningless, but technically possible.
